requirements.txt
streamlit
diffusers
transformers
torch
openai
requests
Pillow
